[ADD] Found the 64-bits big endian file. Fixed last things for this.

+++ src/read_data.c
@@ -33,7 +33,7 @@ uint32_t	read_uint32(uint32_t nb, int endian)
 	return (res);
 }
 
-uint64_t	read_uint64(uint32_t nb, int endian)
+uint64_t	read_uint64(uint64_t nb, int endian)
 {
 	uint64_t	res;
